Reorganized annoy code for greater modularity.
